Fix build in Docker targets
There are some inconveniences when building in container and writing in
shared volume. The biggest of those are that by doing that the resulting
target directory is owned by root and in some rare cases - you end up
with unnecessary files in the target directory. To solve that the
project files are copied in the container and not mounted and the target
directory is copied back to the host when the build finishes.